==> notebooks/00_check_noisy_data.ipynb <==
import pandas as pd
import numpy as np 
import pickle
import os 

observations = {}

for n_var in [3,5,10,20]:
    observations[n_var] = {}
    for std in [0.5,0.75]:
        observations[n_var][std] = {}
        DATA_PATH =  os.path.join('..','data','synthetic',f"data_N{n_var}_std{std}")
        for generative_process in range(1,21):
            with open(os.path.join(DATA_PATH,f"data_{generative_process}.pkl"),'rb') as f:
                observation, _ , _ , _ = pickle.load(f)

            observations[n_var][std][generative_process] = observation

for n_var in [3,5,10,20]:
    for std in [0.5,0.75]:
        for generative_process in range(1,21):
            for multivariate_ts in observations[n_var][std][generative_process]:
                #check for na 
                if multivariate_ts.isna().sum().sum() > 0:
                    print(multivariate_ts.isna().sum())
                    print(f"Found NA in {n_var} {std} {generative_process}")
                #check for inf
                if multivariate_ts.isin([np.inf, -np.inf]).sum().sum() > 0:
                    print(multivariate_ts.isin([np.inf, -np.inf]).sum().sum())
                    print(f"Found inf in {n_var} {std} {generative_process}")
