==> notebooks/06_compute_metrics.ipynb <==
import os
import sys
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns
import pickle
import pandas as pd
from sklearn.metrics import average_precision_score, accuracy_score, precision_score, recall_score, f1_score, balanced_accuracy_score, roc_auc_score, precision_recall_curve, roc_curve, auc

sys.path.append('..')
from src.benchmark.var import VAR
from src.benchmark.d2c_wrapper import D2C
from src.benchmark.metrics import make_plots, compute_roc_auc_curves, precision_top_k
from src.descriptors.d2c_past_gen import DescriptorsGenerator

results = {}
pickle_path = os.path.join('..','results', 'predictions')
methods = ['d2c','var','dynotears','granger','pcmci','varlingam']
variables = [3,5,10,20]
noises = [0.01,0.1,0.3,0.5,0.75]

for method in methods:
    for n_variables in variables:
        for noise_std in noises:

            df = pd.read_pickle(os.path.join(pickle_path,method,f'n{n_variables}_s{noise_std}.pkl'))

            num_class_1 = int(df['is_causal'].sum())
            num_class_0 = int(len(df) - num_class_1)

            p10 = precision_top_k(df,top_k=10)
            p50 = precision_top_k(df,top_k=50)
            p100 = precision_top_k(df,top_k=100)
            p500 = precision_top_k(df,top_k=500)
            p1000 = precision_top_k(df,top_k=1000)

            y_test = df['is_causal']
            y_score = df['predicted_proba']
            y_hat = y_score > 0.5

            accuracy = accuracy_score(y_test, y_hat)
            precision = precision_score(y_test, y_hat, zero_division=np.nan)
            recall = recall_score(y_test, y_hat, zero_division=np.nan)
            f1 = f1_score(y_test, y_hat, zero_division=np.nan)
            ber = 1 - balanced_accuracy_score(y_test, y_hat)
            roc_auc = roc_auc_score(y_test, y_score)
            p_auc_10 = roc_auc_score(y_test, y_score, max_fpr=0.1)
            p_auc_25 = roc_auc_score(y_test, y_score, max_fpr=0.25)
            p_auc_50 = roc_auc_score(y_test, y_score, max_fpr=0.50)

            
            precisions, recalls, _ = precision_recall_curve(y_test, y_score)
            pr_auc = auc(recalls, precisions)
            
            ap = average_precision_score(y_test, y_score)

            results[(method,n_variables,noise_std)] = {'n1':num_class_1,
                                                       'n0':num_class_0,
                                                       'p10':p10, 
                                                       'p50':p50, 
                                                       'p100':p100, 
                                                       'p500':p500, 
                                                       'p1000':p1000, 
                                                       'accuracy':accuracy, 
                                                       'precision':precision, 
                                                       'recall':recall, 
                                                       'f1':f1, 
                                                       'ber':ber, 
                                                       'roc_auc':roc_auc, 
                                                       'p_auc_10':p_auc_10,
                                                       'p_auc_25':p_auc_25,
                                                       'p_auc_50':p_auc_50,
                                                       'pr_auc':pr_auc, 
                                                       'ap':ap
                                                       }

results = pd.DataFrame(results,columns=[(method,n_variables,noise_std) for method in methods for n_variables in variables for noise_std in noises]).T
results.index = pd.MultiIndex.from_tuples(results.index, names=['method','n_variables','noise_std'])
       
    

results.n1 = results.n1.astype(int)
results.n0 = results.n0.astype(int)

results.head()

results.to_csv(os.path.join('..','results','metrics','metrics.csv'))

Computing metrics per each generative process

GRAPHS_PER_PROCESS = 50
results = {}
pickle_path = os.path.join('..','results', 'predictions')
methods = ['d2c','var','dynotears','granger','pcmci','varlingam']
variables = [3,5,10,20]
noises = [0.01,0.1,0.3,0.5,0.75]

for method in methods:
    for n_variables in variables:
        for noise_std in noises:

            df = pd.read_pickle(os.path.join(pickle_path,method,f'n{n_variables}_s{noise_std}.pkl'))

            df['generative_process'] = df['graph_id'] // GRAPHS_PER_PROCESS

            for process in df['generative_process'].unique():
                df_process = df[df['generative_process'] == process]
                
                num_class_1 = int(df_process['is_causal'].sum())
                num_class_0 = int(len(df_process) - num_class_1)

                p10 = precision_top_k(df_process,top_k=10)
                p50 = precision_top_k(df_process,top_k=50)
                p100 = precision_top_k(df_process,top_k=100)
                p500 = precision_top_k(df_process,top_k=500)
                p1000 = precision_top_k(df_process,top_k=1000)

                y_test = df_process['is_causal']
                y_score = df_process['predicted_proba']
                y_hat = y_score > 0.5

                accuracy = accuracy_score(y_test, y_hat)
                precision = precision_score(y_test, y_hat, zero_division=np.nan)
                recall = recall_score(y_test, y_hat, zero_division=np.nan)
                f1 = f1_score(y_test, y_hat, zero_division=np.nan)
                ber = 1 - balanced_accuracy_score(y_test, y_hat)
                roc_auc = roc_auc_score(y_test, y_score)
                p_auc_10 = roc_auc_score(y_test, y_score, max_fpr=0.1)
                p_auc_25 = roc_auc_score(y_test, y_score, max_fpr=0.25)
                p_auc_50 = roc_auc_score(y_test, y_score, max_fpr=0.50)

                
                precisions, recalls, _ = precision_recall_curve(y_test, y_score)
                pr_auc = auc(recalls, precisions)
                
                ap = average_precision_score(y_test, y_score)

                results[(process,method,n_variables,noise_std)] = {'n1':num_class_1,
                                                                    'n0':num_class_0,
                                                                    'p10':p10, 
                                                                    'p50':p50, 
                                                                    'p100':p100, 
                                                                    'p500':p500, 
                                                                    'p1000':p1000, 
                                                                    'accuracy':accuracy, 
                                                                    'precision':precision, 
                                                                    'recall':recall, 
                                                                    'f1':f1, 
                                                                    'ber':ber, 
                                                                    'roc_auc':roc_auc, 
                                                                    'p_auc_10':p_auc_10,
                                                                    'p_auc_25':p_auc_25,
                                                                    'p_auc_50':p_auc_50,
                                                                    'pr_auc':pr_auc, 
                                                                    'ap':ap
                                                                    }

results = pd.DataFrame(results).T
results.index = pd.MultiIndex.from_tuples(results.index, names=['process','method','n_variables','noise_std'])
       
    

results

results.to_csv(os.path.join('..','results','metrics','metrics_per_process.csv'))

import pandas as pd 
import os
# Rounding
#open metrics.csv
metrics = pd.read_csv(os.path.join('..','results','metrics','metrics.csv'),index_col=[0,1,2])
metrics = metrics.round(2)
metrics.to_csv(os.path.join('..','results','metrics','metrics_rounded.csv'))
